+++ crates/testkit/src/testers/data/actor.rs
@@ -80,6 +80,7 @@ impl DataActor for DataTester {
                     None,
                     None,
                     client_id,
+                    None,
                     request_params.clone(),
                 );
             }