+++ docs/how_to/loading_external_data.py
@@ -29,6 +29,7 @@ from nautilus_trader.model import AccountType
 from nautilus_trader.model import Currency
 from nautilus_trader.model import OmsType
 from nautilus_trader.model import Quantity
+from nautilus_trader.model import NautilusDataType
 from nautilus_trader.persistence import ParquetDataCatalog
 from nautilus_trader.testkit.providers import TestDataProvider
 from nautilus_trader.testkit.providers import TestInstrumentProvider
@@ -131,7 +132,7 @@ venue_configs = [
 data_configs = [
     BacktestDataConfig(
         catalog_path=str(CATALOG_PATH),
-        data_type="QuoteTick",
+        data_type=NautilusDataType.QuoteTick,
         instrument_id=instrument.id,
         start_time=start,
         end_time=end,