+++ docs/tutorials/backtest_orderbook_bybit.py
@@ -86,6 +86,7 @@ from nautilus_trader.model import (
     CryptoPerpetual,
     Currency,
     InstrumentId,
+    NautilusDataType,
     OmsType,
     Price,
     Quantity,
@@ -185,7 +186,7 @@ book_type = BookType.L2_MBP
 data_configs = [
     BacktestDataConfig(
         catalog_path=str(CATALOG_PATH),
-        data_type="OrderBookDelta",
+        data_type=NautilusDataType.OrderBookDelta,
         instrument_id=instrument.id,
     ),
 ]